+++ GameProject/src/main/java/com/example/gameproject/Map2Controller.java
@@ -162,21 +162,21 @@ public class Map2Controller implements Initializable {
         towersPosition.add(p3);
 
         ArrayList<Wave> attackWaves = new ArrayList<>();
-        ShieldRaider shieldRaider1 = new ShieldRaider();
-        WizardRaider shieldRaider2 = new WizardRaider();
-        ShieldRaider shieldRaider3 = new ShieldRaider();
-        SpeedyRaider shieldRaider4 = new SpeedyRaider();
-        WizardRaider shieldRaider5 = new WizardRaider();
-        SpeedyRaider shieldRaider6 = new SpeedyRaider();
-        ShieldRaider shieldRaider7 = new ShieldRaider();
+        ShieldRaider raider1 = new ShieldRaider();
+        SpeedyRaider raider2 = new SpeedyRaider();
+        WizardRaider raider3 = new WizardRaider();
+        SpeedyRaider raider4 = new SpeedyRaider();
+        WizardRaider raider5 = new WizardRaider();
+        SpeedyRaider raider6 = new SpeedyRaider();
+        ShieldRaider raider7 = new ShieldRaider();
 
-        Wave wave1 = new Wave(shieldRaider1, 3);
-        Wave wave2 = new Wave(shieldRaider2, 6);
-        Wave wave3 = new Wave(shieldRaider3, 8);
-        Wave wave4 = new Wave(shieldRaider4, 10);
-        Wave wave5 = new Wave(shieldRaider5, 10);
-        Wave wave6 = new Wave(shieldRaider4, 10);
-        Wave wave7 = new Wave(shieldRaider5, 13);
+        Wave wave1 = new Wave(raider1, 3);
+        Wave wave2 = new Wave(raider2, 6);
+        Wave wave3 = new Wave(raider3, 8);
+        Wave wave4 = new Wave(raider4, 10);
+        Wave wave5 = new Wave(raider5, 10);
+        Wave wave6 = new Wave(raider6, 10);
+        Wave wave7 = new Wave(raider7, 13);
 
         attackWaves.add(wave1);
         attackWaves.add(wave2);
@@ -235,14 +235,14 @@ public class Map2Controller implements Initializable {
 
         if (SpellsController.getInstance().drop()) {
             MapController.getInstance().bombAttacks(pane);
-            for(Raider raider:MapController.map.getAliveRaiders()){
-                int index=MapController.map.getAliveRaiders().indexOf(raider);
+            for(Raider raider:MapController.getMap().getAliveRaiders()){
+                int index=MapController.getMap().getAliveRaiders().indexOf(raider);
                 VBox vBox=vboxesList.get(index);
-                PathTransition pathTransition=MapController.map.getPathTransitions().get(index);
+                PathTransition pathTransition=MapController.getMap().getPathTransitions().get(index);
                 removeRaider(raider, vBox, pathTransition);
             }
 
-            MapController.map.getAliveRaiders().clear();
+            MapController.getMap().getAliveRaiders().clear();
             setSpellCounts();
 
         }
@@ -276,6 +276,9 @@ public class Map2Controller implements Initializable {
 
     @FXML
     void dropHealth(MouseEvent event) throws URISyntaxException {
+        if(PlayerController.getPlayer().getHealth()==20){
+            return;
+        }
         HealthSpell healthSpell = new HealthSpell();
         SpellsController.setSpell(healthSpell);
         if (SpellsController.getInstance().drop()) {
@@ -471,7 +474,7 @@ public class Map2Controller implements Initializable {
             if (isFinished) {
                 return;
             }
-            if(!MapController.map.getAliveRaiders().contains(currentRaider)) {
+            if(!MapController.getMap().getAliveRaiders().contains(currentRaider)) {
                 return;
             }
             currentRaider.setHealth(health);
@@ -483,7 +486,7 @@ public class Map2Controller implements Initializable {
                         MapController.getInstance().archerTowerAttack(tower, currentRaider, point, vBox, pane);
 
                     } else if (tower instanceof Artillery && (!(currentRaider instanceof FlyerRaider))) {
-                        MapController.getInstance().artilleryTowerAttack(tower, currentRaider, point, vBox, pane);
+                        MapController.getInstance().artilleryTowerAttack( point, vBox, pane);
                         attackNearRaiders(tower,vBox,((Artillery) tower).getDamageRange());
                         return;
 
@@ -499,7 +502,7 @@ public class Map2Controller implements Initializable {
                 }
                 if(currentRaider.getHealth()<=0){
                     removeRaider(currentRaider, vBox, pathTransition);
-                    MapController.map.getAliveRaiders().remove(currentRaider);
+                    MapController.getMap().getAliveRaiders().remove(currentRaider);
                     return;
                 }
             }
@@ -511,12 +514,12 @@ public class Map2Controller implements Initializable {
 
     public  void attackNearRaiders(Tower tower,VBox vBox1,int range){
         ArrayList<Integer> indexes=new ArrayList<>();
-        for(Raider raider:MapController.map.getAliveRaiders()){
-            int index=MapController.map.getAliveRaiders().indexOf(raider);
+        for(Raider raider:MapController.getMap().getAliveRaiders()){
+            int index=MapController.getMap().getAliveRaiders().indexOf(raider);
             VBox vBox2=vboxesList.get(index);
             double distance=Math.hypot(vBox1.getTranslateX()-vBox2.getTranslateX(),vBox1.getTranslateY()-vBox2.getTranslateY());
             if(distance<range) {
-                PathTransition pathTransition = MapController.map.getPathTransitions().get(index);
+                PathTransition pathTransition = MapController.getMap().getPathTransitions().get(index);
                 raider.setHealth(raider.getHealth()-tower.getDestroyPower());
 
                 if(raider.getHealth()<=0) {
@@ -529,7 +532,7 @@ public class Map2Controller implements Initializable {
             }
         }
         for(int index:indexes){
-            MapController.map.getAliveRaiders().remove(index);
+            MapController.getMap().getAliveRaiders().remove(index);
         }
     }
 
@@ -612,6 +615,7 @@ public class Map2Controller implements Initializable {
         UpgradeBox.setVisible(false);
         Tower selectedTower = MapController.getInstance().getTower(newPath);
         if(MapController.getInstance().checkTowerLevelForUpgrade(selectedTower)){
+            spellsBox.setVisible(true);
             return;
         }
         if (checkCoins(selectedTower)) {
@@ -635,6 +639,7 @@ public class Map2Controller implements Initializable {
             PageController.showAlert("Error", "You don't have enough coins to upgrade this tower!!", "", Alert.AlertType.ERROR);
             UpgradeBox.setVisible(false);
             towersBox.setVisible(false);
+            spellsBox.setVisible(true);
             return true;
         }
         return false;
